feat: ORM: add APIs for defining and bootstrapping new db (#82)
This PR introduces a set of new APIs for further deterministically defining and bootstrapping new database:
1. `orm_bootstrap_table_name`: pre-define the table_name.
2. `orm_bootstrap_create_table_params`: pre-define the table creat statement.
3. `orm_bootstrap_indexes_params`: optional list of statement for creating indexes.

For 2 and 3, we can also provide typeddict of params for generating query statement with table_create_stmt and table_create_index_stmt correspondingly.

With this PR, the previous way to pre-define table_name by setting `_orm_table_name` is deprecated in flavor of `orm_bootstrap_table_name`, `_orm_table_name` now is an internal attr and should not be directly assigned.

Other change:
1. utils: fix check_db_integrity when checking against specific table.
2. test_ci: use SonarSource/sonarqube-scan-action@v4 as the previous one is deprecated by sonarcloud.