fix: table_spec: fix insert query composing doesn't include cols definition (#100)
Previously, when generating insert query, the statement of specifying which cols will be assigned are not included, but only the values to insert are specified. 

This will not cause problem when we do insert with row that has all cols assigned, and with the same schema as table, but will cause problem when we insert with row that cols order dosen't match schema, or cols are partially set(like without specifying `rowid`).

This PR fixes the above issue, and add tests to cover the above cases.