ci: add test base images and compose file for testing on supported ubuntu version  (#127)
This PR introduces the Dockerfile for generating test base container image for running tests on supported Ubuntu versions(from 18.04 to 24.04). A compose file is also added for quick setup and running the tests.

For the docker compose file, it sets up the container as follow:
1. mount project root as read-only into the container.
2. mount `<project_root>/.test_result` as read-write into container for holding the test results.

No code or feature changes are introduced.

Other changes:
1. devcontainer.json: add docker-in-docker feature into the dev container.
3. tests.conftest: rename `SQLITE3_COMPILE_OPTION_FLAGS` to `SQLITE3_FEATURE_FLAGS`.
4. tests.conftest: add `STRICT_AVAILABLE` flag to `SQLITE3_FEATURE_FLAGS`.